requirements.txt
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.5.2
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
GraphRicciCurvature==0.5.3.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.41.1
matplotlib==3.7.5
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
networkit==11.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai==1.52.2
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pinecone==5.3.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
POT==0.9.4
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-louvain==0.16
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.1
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
umap==0.1.1
umap-learn==0.5.7
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
